q1/hr-onboarding-assistant/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
PyMuPDF==1.23.7
python-docx==1.0.1
langchain==0.0.350
sentence-transformers==2.2.2
qdrant-client==1.7.0
google-generativeai==0.3.2 
huggingface-hub==0.16.4